// frontend/src/api/generated/model/accountOut.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Family Finance Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { AccountOutDefaultOwnerId } from './accountOutDefaultOwnerId';

export interface AccountOut {
  id: number;
  name: string;
  last4: string;
  account_kind: string;
  default_owner_id: AccountOutDefaultOwnerId;
  source_format: string;
}